+++ src/app/services/sdk/server/socket/index.tsx
@@ -10,7 +10,13 @@
 import {base64ToU8a, uint8ToBase64} from '../../fileManager/http/utils';
 import {IServerRequest, serverKeys} from '../index';
 import ElectronService from '../../../electron';
-import {EventWasmInit, EventWebSocketClose, EventWebSocketOpen} from "../../../events";
+import {
+    EventCheckNetwork,
+    EventNetworkStatus,
+    EventWasmInit,
+    EventWebSocketClose,
+    EventWebSocketOpen
+} from "../../../events";
 import {C_LOCALSTORAGE} from "../../const";
 
 export const defaultGateway = 'cyrus.river.im';
@@ -71,7 +77,7 @@ export default class Socket {
 
         window.addEventListener('online', () => {
             this.online = true;
-            this.dispatchEvent('networkStatus', {online: true});
+            this.dispatchEvent(EventNetworkStatus, {online: true});
             this.initTimeout = setTimeout(() => {
                 this.initWebSocket();
             }, 10);
@@ -79,14 +85,14 @@ export default class Socket {
 
         window.addEventListener('offline', () => {
             this.online = false;
-            this.dispatchEvent('networkStatus', {online: false});
+            this.dispatchEvent(EventNetworkStatus, {online: false});
             this.closeWire();
         });
 
         setTimeout(() => {
             if (!navigator.onLine) {
                 this.online = false;
-                this.dispatchEvent('networkStatus', {online: false});
+                this.dispatchEvent(EventNetworkStatus, {online: false});
                 this.closeWire();
             }
         }, 3000);
@@ -96,6 +102,10 @@ export default class Socket {
         if (connection) {
             connection.addEventListener('change', () => {
                 window.console.log('connection changed!');
+                if (navigator.onLine !== undefined && this.online !== navigator.onLine) {
+                    this.online = navigator.onLine;
+                    this.dispatchEvent(EventNetworkStatus, {online: navigator.onLine});
+                }
                 if (this.online) {
                     this.closeWire();
                 }
@@ -136,9 +146,7 @@ export default class Socket {
                 case 'wsSend':
                     if (this.socket && this.connected && this.socket.readyState === WebSocket.OPEN) {
                         this.socket.send(base64ToU8a(d.data));
-                        if (this.lastReceiveTime >= this.lastSendTime) {
-                            this.lastSendTime = Date.now();
-                        }
+                        this.lastSendTime = Date.now();
                     }
                     break;
                 case 'wsError':
@@ -299,7 +307,7 @@ export default class Socket {
             this.tryCounter = 0;
             this.connected = true;
             this.lastSendTime = Date.now();
-            this.lastReceiveTime = Date.now();
+            this.lastReceiveTime = Date.now() + 1;
             if (this.started) {
                 this.dispatchEvent(EventWebSocketOpen, null);
             }
@@ -375,8 +383,7 @@ export default class Socket {
             if (this.lastSendTime > this.lastReceiveTime && this.online) {
                 const now = Date.now();
                 if (now - this.lastSendTime > 12000) {
-                    window.console.log('bad network');
-                    this.closeWire();
+                    this.dispatchEvent(EventCheckNetwork, {});
                 }
             }
         }, 12000);